refine gpu_mem_info api

+++ src/cpp_worker/adapter-llama.cpp
@@ -1,12 +1,12 @@
 #include "adapter-llama.hpp"
 #include "logging.hpp"
 extern "C" {
-void get_gpu_mem_info(size_t *free_byte, size_t *total_byte) {
+void get_gpu_mem_info(uint64_t *free_byte, uint64_t *total_byte) {
   CUDA_CALL(cudaMemGetInfo(free_byte, total_byte));
 }
 
 void log_gpu_mem_info() {
-  size_t free_byte, total_byte;
+  uint64_t free_byte, total_byte;
   get_gpu_mem_info(&free_byte, &total_byte);
   LOG(ERROR) << "GPU memory: free " << free_byte/1024.0/1024.0/1024.0 << "GB, total " << total_byte/1024.0/1024.0/1024.0 << "GB";
 }
