+++ src/cpp_worker/adapter-llama.hpp
@@ -3,6 +3,6 @@
 #include "profiler.hpp"
 
 extern "C" {
-  void get_gpu_mem_info(size_t* free, size_t* total);
+  void get_gpu_mem_info(uint64_t* free, uint64_t* total);
   void log_gpu_mem_info();
 }
